tty.sys: new service to redirect stdin/stdout to serial ports - stdin only implemented
stdout implementation not done yet

+++ msx/apps/libraries/extbio/serial-helper-read-char.c
@@ -7,8 +7,8 @@
 extern uint8_t port_number;
 
 uint8_t serial_read_char(uint8_t *const c) __sdcccall(1) {
-  uint16_t       size   = 1;
-  const uint16_t result = serial_demand_read(port_number, c, &size, 1000);
+  uint16_t      size   = 1;
+  const uint8_t result = serial_demand_read(port_number, c, &size, 1000);
   if (result)
     return result;
